Reorganize client/common packages
