+++ builder/azure/common/client/detect_azure.go
@@ -1,6 +1,6 @@
 // +build !linux
 
-package common
+package client
 
 // IsAzure returns true if Packer is running on Azure (currently only works on Linux)
 func IsAzure() bool {
