+++ builder/azure/common/client/detect_azure_linux.go
@@ -1,4 +1,4 @@
-package common
+package client
 
 import (
 	"bytes"
