+++ builder/azure/common/client/detect_azure_linux_test.go
@@ -1,4 +1,4 @@
-package common
+package client
 
 import (
 	"io/ioutil"
