+++ builder/azure/common/client/devicelogin.go
@@ -1,4 +1,4 @@
-package common
+package client
 
 import (
 	"context"
